Reorganized files to different folders for faster file locating in qt creator. Decided to use three folders: audio, gui(not ui so it wont get confused with ui_ stuff with headers) and video.
# Conflicts:
#	HEVC_conference.pro
#	src/stun.cpp
#	src/stun.h

+++ src/gui/conferenceview.cpp
@@ -46,8 +46,8 @@ void ConferenceView::callingTo(QString callID, QString name)
 void ConferenceView::addWidgetToLayout(CallState state, QWidget* widget, QString name, QString callID)
 {
   locMutex_.lock();
-  int row = row_;
-  int column = column_;
+  uint8_t row = row_;
+  uint8_t column = column_;
 
   if(!freedLocs_.empty())
   {
